requirements.txt
python-dotenv
slack-bolt
supabase
streamlit
